refactor: auth is a page
