+++ frontend/svelte/src/AuthPage.svelte
